+++ include/linux/mfd/samsung/core.h
@@ -122,8 +122,8 @@ extern int sec_rtc_bulk_read(struct sec_pmic_dev *sec_pmic, u8 reg, int count,
 extern int sec_rtc_write(struct sec_pmic_dev *sec_pmic, u8 reg, u8 value);
 extern int sec_rtc_bulk_write(struct sec_pmic_dev *sec_pmic, u8 reg, int count,
 				u8 *buf);
-extern int sec_rtc_update(struct sec_pmic_dev *sec_pmic, u8 reg, u8 val,
-				u8 mask);
+extern int sec_rtc_update(struct sec_pmic_dev *sec_pmic, u8 reg, unsigned int val,
+				unsigned int mask);
 
 /**
  * sec_regulator_data - regulator data
